move common fictures to conftest
